+++ api/src/main/java/de/verdox/mccreativelab/wrapper/platform/data/MCCDataPackInterceptor.java
@@ -1,4 +1,4 @@
-package de.verdox.mccreativelab.wrapper.data;
+package de.verdox.mccreativelab.wrapper.platform.data;
 
 import com.google.gson.JsonObject;
 import net.kyori.adventure.key.Key;
